♻️ hotelId 패치 커스텀훅으로 리팩터링
